requirements.txt
streamlit
chromadb
langchain-community
langchain-ollama
sentence-transformers
pypdf
pytube
openai-whisper
python-dotenv
# Nowe zależności dla Refaktoryzacji:
rapidfuzz==3.6.1       # Dla Punktu 5: Obsidian Linking (Fuzzy match)
pdfplumber==0.10.3     # Dla Punktu 6: Lepsza ekstrakcja tabel z PDF
watchdog==3.0.0        # Dla Punktu 2: Incremental Indexing (monitoring plików)
pyannote.audio         # Dla Diaryzacji mówców
faster-whisper         # Szybszy Whisper
yt-dlp                 # Pobieranie z YT